% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run.R
\name{as.camodel_initmat}
\alias{as.camodel_initmat}
\title{Convert a matrix to a CA model landscape}
\usage{
as.camodel_initmat(m, levels = NULL)
}
\arguments{
\item{m}{The input matrix (numeric, character or factor)}

\item{levels}{The levels to use in the resulting landscape. If \code{NULL}, the unique
values of the input matrix are used as levels. Set this manually if you want the 
resulting landscape to have extra levels that are not present in the original matrix.}
}
\description{
Convert a matrix to a CA model landscape for later use with
 \link{run_camodel} or \link{run_meanfield}.
}
\examples{

# Simple conversion of a matrix with regular patterns
x <- seq(0, 2 * pi, l = 256) 
z <- outer(x, x, function(x, y) as.numeric(sin(10*x) + cos(10*y) > 0.8))
mat <- as.camodel_initmat(z)
summary(mat)
image(mat)

# This is a character matrix. We need to convert it to use it as input to 
# run_camodel()
size <- 64
m <- matrix(ifelse(runif(size^2) < .5, "TREE", "EMPTY"), nrow = size, ncol = size)
m <- as.camodel_initmat(m)
summary(m) # this is a landscape object 
image(m)

# Start a simulation using this matrix 
mod <- ca_library("forestgap")
out <- run_camodel(mod, m, seq(0, 256))
plot(out)


}
\seealso{
generate_initmat, run_camodel, run_meanfield
}
