% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ca_library.R
\name{ca_library}
\alias{ca_library}
\title{Library of probabilistic cellular automata}
\usage{
ca_library(model, parms = NULL, neighbors = NULL, wrap = TRUE)
}
\arguments{
\item{model}{The model to return, as a string. See Details for the full list of models
included with \code{chouca}.}

\item{parms}{The model parameters to use, as a named list. If unset, the model default
parameters will be used.}

\item{neighbors}{The number of neighbors to use in the cellular automaton (4 for 4-way 
or von-Neumann neghborhood, or 8 for a Moore neighborhood). If unset, the model 
default neighborhood will be used.}

\item{wrap}{Whether the 2D grid should wrap around at the edges.}
}
\description{
Get one of the PCA model included in \code{chouca}
}
\examples{

# Import a model, create an initial landscape and run it for ten iterations
forestgap_model <- ca_library("forestgap")
im <- generate_initmat(forestgap_model, c(0.5, 0.5), nr = 64, nc = 100) 
run_camodel(forestgap_model, im, niter = 10) 

}
