% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run.R
\name{generate_initmat}
\alias{generate_initmat}
\title{Generate an initial matrix for a \code{chouca} model}
\usage{
generate_initmat(mod, pvec, nr, nc = nr)
}
\arguments{
\item{mod}{A stochastic cellular automaton model created by \code{\link{camodel}}}

\item{pvec}{A numeric vector of covers for each state in the initial configuration.}

\item{nr}{The number of rows of the output matrix}

\item{nc}{The number of columns of the output matrix}
}
\description{
Helper to create a initial landscape (matrix) with specified covers for 
  a cellular automaton
}
\details{
This function is a helper to build a starting configuration (matrix) for a 
    stochastic cellular automaton. Based on the definition of the model and the 
    specified starting covers (in \code{pvec}). It will produce a landscape with 
    expected global cover of each state equal to the covers in \code{pvec}, and a 
    completely random spatial structure. 
  
  The length of the \code{pvec} vector must match the number of possible cell states 
    in the model. If present, the names of \code{pvec} must match the states
    defined in the model. In this case, they will be used to determine which state 
    gets which starting cover. 
  
  The \code{pvec} will be normalized to sum to one, emitting a warning if this 
    produces a meaningful change in covers.
}
\seealso{
as.camodel_initmat
}
