% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parsing.R
\name{camodel}
\alias{camodel}
\title{Definition of a probabilistic cellular automaton}
\usage{
camodel(..., parms, all_states = NULL, verbose = TRUE)
}
\arguments{
\item{...}{a number of transition descriptions, as built by the
\code{\link{transition}} function (see Details)}

\item{parms}{a named list of parameters, which should be all numeric, single values}

\item{all_states}{the complete list of states (a character vector). If unspecified,
it will be guessed from the transition rules, but you should really pass it to
make sure your model definition is correct}

\item{verbose}{whether information should be printed when parsing the model
definition}
}
\description{
High-level definition of a probabilistic cellular automaton
}
